ELE-1411: read zip archives on 64-bit ABIs
The zip header structs declared their 4-byte fields as unsigned long, which is 8 bytes on arm64 and x86_64, so fread(&field, 4, ...) filled only the low half and every signature check compared against uninitialised stack. The end-of-central-directory scan likewise read 8 bytes through an unsigned long cast. Fixed-width fields and a byte comparison make ZLZipArchive::Open succeed on 64-bit, which mounting the APK with AKUMountVirtualDirectory depends on.

Verified on an arm64 API 36 emulator: before the change Open fails on any zip; after it, stored and deflated entries (fully cached and block-streamed, including a backward seek) read back byte-identical, from a test zip and from an installed 294 MB signed APK.

+++ src/zlcore/ZLZipArchive.h
@@ -4,20 +4,22 @@
 #ifndef ZLZIPARCHIVE_H
 #define ZLZIPARCHIVE_H
 
+#include <stdint.h>
+
 //================================================================//
 // ZLZipArchiveHeader
 //================================================================//
 class ZLZipArchiveHeader {
 public:
 
-	unsigned long	mSignature;			// 4 End of central directory signature = 0x06054b50
-	unsigned short	mDiskNumber;		// 2 Number of this disk
-	unsigned short	mStartDisk;			// 2 Disk where central directory starts
-	unsigned short	mTotalDiskEntries;	// 2 Total number of entries on disk
-	unsigned short	mTotalEntries;		// 2 Total number of central in archive
-	unsigned long	mCDSize;			// 4 Size of central directory in bytes
-	unsigned long	mCDAddr;			// 4 Offset of start of central directory, relative to start of archive
-	unsigned short	mCommentLength;		// 2 ZIP file comment length
+	uint32_t	mSignature;			// 4 End of central directory signature = 0x06054b50
+	uint16_t	mDiskNumber;		// 2 Number of this disk
+	uint16_t	mStartDisk;			// 2 Disk where central directory starts
+	uint16_t	mTotalDiskEntries;	// 2 Total number of entries on disk
+	uint16_t	mTotalEntries;		// 2 Total number of central in archive
+	uint32_t	mCDSize;			// 4 Size of central directory in bytes
+	uint32_t	mCDAddr;			// 4 Offset of start of central directory, relative to start of archive
+	uint16_t	mCommentLength;		// 2 ZIP file comment length
 	
 	//----------------------------------------------------------------//
 	int		FindAndRead		( FILE* file );
@@ -29,23 +31,23 @@ public:
 class ZLZipEntryHeader {
 public:
 
-	unsigned long	mSignature;				// 4 Central directory file header signature = 0x02014b50
-	unsigned short	mByVersion;				// 2 Version made by
-	unsigned short	mVersionNeeded;			// 2 Version needed to extract (minimum)
-	unsigned short	mFlag;					// 2 General purpose bit flag
-	unsigned short	mCompression;			// 2 Compression method
-	unsigned short	mLastModTime;			// 2 File last modification time
-	unsigned short	mLastModDate;			// 2 File last modification date
-	unsigned long	mCrc32;					// 4 CRC-32
-	unsigned long	mCompressedSize;		// 4 Compressed size
-	unsigned long	mUncompressedSize;		// 4 Uncompressed size
-	unsigned short	mNameLength;			// 2 File name length (n)
-	unsigned short	mExtraFieldLength;		// 2 Extra field length (m)
-	unsigned short	mCommentLength;			// 2 File comment length (k)
-	unsigned short	mDiskNumber;			// 2 Disk number where file starts
-	unsigned short	mInternalAttributes;	// 2 Internal file attributes
-	unsigned long	mExternalAttributes;	// 4 External file attributes
-	unsigned long	mFileHeaderAddr;		// 4 Relative offset of file header
+	uint32_t	mSignature;				// 4 Central directory file header signature = 0x02014b50
+	uint16_t	mByVersion;				// 2 Version made by
+	uint16_t	mVersionNeeded;			// 2 Version needed to extract (minimum)
+	uint16_t	mFlag;					// 2 General purpose bit flag
+	uint16_t	mCompression;			// 2 Compression method
+	uint16_t	mLastModTime;			// 2 File last modification time
+	uint16_t	mLastModDate;			// 2 File last modification date
+	uint32_t	mCrc32;					// 4 CRC-32
+	uint32_t	mCompressedSize;		// 4 Compressed size
+	uint32_t	mUncompressedSize;		// 4 Uncompressed size
+	uint16_t	mNameLength;			// 2 File name length (n)
+	uint16_t	mExtraFieldLength;		// 2 Extra field length (m)
+	uint16_t	mCommentLength;			// 2 File comment length (k)
+	uint16_t	mDiskNumber;			// 2 Disk number where file starts
+	uint16_t	mInternalAttributes;	// 2 Internal file attributes
+	uint32_t	mExternalAttributes;	// 4 External file attributes
+	uint32_t	mFileHeaderAddr;		// 4 Relative offset of file header
 	
 	//----------------------------------------------------------------//
 	int		Read	( FILE* file );
@@ -57,17 +59,17 @@ public:
 class ZLZipFileHeader {
 public:
 
-	unsigned long	mSignature;				// 4	Local file header signature = 0x04034b50 (read as a little-endian number)
-	unsigned short	mVersionNeeded;			// 2	Version needed to extract (minimum)
-	unsigned short	mFlag;					// 2	General purpose bit flag
-	unsigned short	mCompression;			// 2	Compression method
-	unsigned short	mLastModTime;			// 2	File last modification time
-	unsigned short	mLastModDate;			// 2	File last modification date
-	unsigned long	mCrc32;					// 4	CRC-32 (*not* to be trusted - Android)
-	unsigned long	mCompressedSize;		// 4	Compressed size (*not* to be trusted - Android)
-	unsigned long	mUncompressedSize;		// 4	Uncompressed size (*not* to be trusted - Android)
-	unsigned short	mNameLength;			// 2	File name length
-	unsigned short	mExtraFieldLength;		// 2	Extra field length
+	uint32_t	mSignature;				// 4	Local file header signature = 0x04034b50 (read as a little-endian number)
+	uint16_t	mVersionNeeded;			// 2	Version needed to extract (minimum)
+	uint16_t	mFlag;					// 2	General purpose bit flag
+	uint16_t	mCompression;			// 2	Compression method
+	uint16_t	mLastModTime;			// 2	File last modification time
+	uint16_t	mLastModDate;			// 2	File last modification date
+	uint32_t	mCrc32;					// 4	CRC-32 (*not* to be trusted - Android)
+	uint32_t	mCompressedSize;		// 4	Compressed size (*not* to be trusted - Android)
+	uint32_t	mUncompressedSize;		// 4	Uncompressed size (*not* to be trusted - Android)
+	uint16_t	mNameLength;			// 2	File name length
+	uint16_t	mExtraFieldLength;		// 2	Extra field length
 
 	//----------------------------------------------------------------//
 	int		Read	( FILE* file );
@@ -80,11 +82,11 @@ class ZLZipFileEntry {
 public:
 
 	std::string		mName;
-	unsigned long	mFileHeaderAddr;
-	unsigned long	mCrc32;
-	unsigned short	mCompression;
-	unsigned long	mCompressedSize;
-	unsigned long	mUncompressedSize;
+	uint32_t	mFileHeaderAddr;
+	uint32_t	mCrc32;
+	uint16_t	mCompression;
+	uint32_t	mCompressedSize;
+	uint32_t	mUncompressedSize;
 	
 	ZLZipFileEntry*	mNext;
 };
